moving hybrid losses into loss_funtions file
